+++ src/main/java/racinggame/RacingGameRound.java
@@ -1,5 +1,7 @@
 package racinggame;
 
+import racinggame.util.RandomNumberGenerator;
+
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
