+++ src/main/java/racinggame/util/RacingGameRandomNumber.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.util;
 
 import java.util.Random;
